Split HHWWggTagProducer into HHWWggTagProducer.cc and HHWWggTagProducer.h
